change the rules for chained call/sel; new tests
